TimeStamp tests and comments.
