+++ src/HdrHistogram.UnitTests/HdrHistogram.UnitTests.csproj
@@ -69,6 +69,7 @@
     <Compile Include="Recording\RecorderTestWithLongHistogram.cs" />
     <Compile Include="Recording\RecorderTestWithLShortHistogram.cs" />
     <Compile Include="ShortHistogramTests.cs" />
+    <Compile Include="TimeStampTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
